Finished implementing elevation niceties

+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/utils/AndroidMath.java
@@ -8,7 +8,7 @@ import android.util.TypedValue;
  */
 
 public class AndroidMath {
-    public static float dp2px(int dp, Resources r) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, r.getDisplayMetrics());
+    public static int dp2px(int dp, Resources r) {
+        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, r.getDisplayMetrics());
     }
 }
